Permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,12 +3,21 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils')
 
 const Mutations = {
     async createItem(parent, args, ctx, info) {
-        // TODO: Check if they are logged in
+        if(!ctx.request.userId) {
+            throw new Error('You must be logged in');
+        }
         const item = await ctx.db.mutation.createItem({
             data: {
+                //This is the way we create relationships in prisma
+                user: {
+                    connect: {
+                        id: ctx.request.userId
+                    }
+                },
                 ...args
             }
         }, info);
@@ -25,15 +34,21 @@ const Mutations = {
             },
         }, info);   
     },
-    deleteItem(parent, args, ctx, info) {
-        const where = {id: args.id };
+    async deleteItem(parent, args, ctx, info) {
+        const where = { id: args.id };
         //1. find item 2. check if they own item 3. Delete it
-        const item = ctx.db.query.item({where}, `{ id title}`);
+        const item = await ctx.db.query.item({where}, `{ id title user {id} }`);
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some
+                (permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+        if(!ownsItem && hasPermissions) {
+            throw new Error('You cannot perform that action')
+        }
         return ctx.db.mutation.deleteItem({ where }, info)
     },
     async signup(parent, args, ctx, info) {
         args.email = args.email.toLowerCase();
-        const password = await bcryp.hash(args.password, 10);
+        const password = await bcrypt.hash(args.password, 10);
         const user = await ctx.db.mutation.createUser({
             data: {
                 ...args,
@@ -129,6 +144,28 @@ const Mutations = {
         })
         return updateUser;
     },
+    async updatePermissions(parent, args, ctx, info) {
+        if (!ctx.request.userId) {
+            throw new Error('You must be logged in');
+        }
+    const currentUser = await ctx.db.query.user({
+                where: {
+                    id: ctx.request.userId,
+                },
+            }, info);
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    return ctx.db.mutation.updateUser({
+        data: {
+            permissions: {
+                // We use the set since we are using an enum in the datmodel
+                set: args.permissions
+            },
+        },
+        where: {
+            id: args.userId
+        }
+    }, info);
+    }
 };
 
 module.exports = Mutations;
